backend/requirements.txt
django==4.2.10
djangorestframework
djangorestframework-simplejwt
django-cors-headers
celery
django-celery-results
psycopg2-binary
pymongo
django-mongoengine
mongoengine
django-rest-framework-mongoengine
django-jalali
pyotp
pillow